basic functionality for POS complete.
item adding/ updating, cart adding/updating/checkout done.

+++ TodoApi/Models/UserProfile.cs
@@ -5,6 +5,6 @@ namespace TodoApi.Models
         public string Name { get; set; }
         public AccessRights rights { get; set; }
 
-        public uint wallet { get; set; }
+        public decimal wallet { get; set; }
     }
 }
